knnlm/requirements.txt
torch>=1.9.0
transformers
datasets
tqdm
scipy
faiss-gpu==1.7.2
sacrebleu
lightning
bitsandbytes
deepspeed
peft
